Change config.json into .env

+++ back/tests/jestGlobalTeardown.js
@@ -1,13 +1,15 @@
 const fs = require("fs");
 const dbName = require("./databaseName");
 const executeQuery = require("../functions/dataBase/executeQuery").run;
+require("dotenv").config();
 
 module.exports = async () => {
+  if (!process.env.DB_DATABASE.endsWith("test")) return;
+
   const connection = await require("../functions/dataBase/createConnection").getDb();
   await executeQuery(connection, "DROP DATABASE ??", [dbName]);
 
   connection.end();
-  fs.writeFileSync(__dirname + "/../config.json", await require(__dirname + "/jestGlobalSetup.js").getRealConfig());
 
   //Remove all STL files for tests
   await new Promise((resolve) => {
